Break out filter testing
